update oid to long in OrganizationDatabase

+++ api/src/main/kotlin/com/openlattice/organizations/OrganizationDatabase.kt
@@ -1,6 +1,6 @@
 package com.openlattice.organizations
 
 data class OrganizationDatabase(
-        val oid: Int,
-        var name: String
+    val oid: Long,
+    var name: String
 )
